Add management_pack rights, right migrations
The management pack rights are bog-standard nacoma rights. The migration
script merely takes an array of something to base the new rights off of,
and adds the new rights accordingly - in this case, anything you can do
for hostgroups, you can - on upgrade - do on a management pack, unless
someone configures things differently.

(This commit is imported from op5lib)

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>