Parser generator: handle error recovery in JS parser.
This makes it possible to still show custom column and column selection, even
though some custom column definition has syntax errors in the definition.

Also the error messages (when using error handling) is improved, because the
incorrect part of the source can be extracted, and also in certain cases, the
relevant preceeding tokens.

(This commit is imported from op5lib)

Signed-off-by: Max Sikström <msikstrom@op5.com>