Custom Columns: Make grammar and visitor support error recovery
The parser generator will soon have support for error recovery in the
javascript parser. But to handle error recovery, error handling productions
needs to be added to the grammar.

This patch adds those productions, and handlers for those productions. But
noone of those productions should be called before error handling is
implemented, because the magic token "error" should never be produced by the
lexer, and therefore no reductions of those productions will occur.

When the improved parser generator is available, syntax errors within a column
should never affect other columns definitions, but just output a syntax error
message on that column, where the error string is outputted

Signed-off-by: Max Sikström <msikstrom@op5.com>