ListView: Don't display the filter when no matching entries
Showing the filter when no matching elements is found is good in many cases
when searching and writing queries, so the user can understand why no matching
entries is found.

That applies to theory. But most users are clicking "unacknowledged problems",
and are expecting the list to be empty, and not three lines of a strange query.
The want the information that everything is ok (no matching entries for
filter), and nothing more.

So if nothing is matched, click on the filter editor button instead, or if in a
widget, show in listview first.

Signed-off-by: Max Sikström <msikstrom@op5.com>