ListView/Multiaction: Don't select multiaction through dropdown in dropdown
Multiaction has evolved. Started as a dropdown above the table. The table was
converted to be fullscreen, the multi action was moved to a toolbox, the
toolbox was split out to a seperate toolbox containing only a multiaction. The
multiaction was lonely in the toolbox, and kept it's options as a dropdown in
the dropdown toolbox.

But the user doesn't like to click that much, so a dropdown in a dropdown is
just a click to much. Break out the multi action to a list of links instead.

Resolves #7274

Signed-off-by: Max Sikström <msikstrom@op5.com>