ORM: Merge code for filter node "or" and "and", and use class inheritance
The code for And and Or operators in the ORM filter structure was exactly the
same, with the only difference in the visitor method

Use late static bindings to reference to the class itself, and then use class
inheritance to reuse the code. This makes the handling of simplifications and
similar methods much easier.

Signed-off-by: Max Sikström <msikstrom@op5.com>