doxygen: Generate documentation for widgets
Apparently, we never did. The listview widget, living in its own module,
had documentation generated, which wouldn't work as well as it should,
because the widget_Base wasn't included in the doc build.

Most of this diff is deleting phpdoc(?) tags from the widgets that were
copy-pasted all over. Now, I'll go create an air field out of bamboo to
get the airplane gods to return, because cargo cult coding is what we do
here.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>