configs: Make cuke nagios config less annoying
Don't log status information, either to status.log or status.sav - we
want to use nagios as a fixture, and we don't care about this.

Log less when we're not in verbose mode - this should still log startups
and shutdowns, right?

Log to /mnt/logs so the log will be available from the buildbot UI
afterwards, because I'm tired of digging for it.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>