Acknowledgements: Make the sticky option super-sticky by default
A sticky option of 1 means "normal sticky", which means "let the
ack remain as long as the state doesn't change" (which is stupid,
since that's what "0" should mean, really), whereas 2 means "let
the ack remain until we recover to an OK or UP state", which is
what we would normally want for acknowledgements.

This patch changes the defaults so sticky acknowledgements are
supersticky by default, and thereby fixing issue #7546.

Signed-off-by: Andreas Ericsson <ae@op5.se>