view/template: Enable custom header and optional loading of dojo
Since the wizard will be using bootstrap as css framework the menu gets warped.
Additionally if the menu is not loaded the dojo scripts will produce a lot of
javascript errors when elements do not exist on the page. A custom header can be
injected using the template variable custom_header.

Signed-off-by: Henrik Smederod <hsmederod@op5.com>