cucumber: Fix summary tests involving counting
The numbers we claimed to be expecting had no relation to the proper
output at all - I think we were matching the current time in the custom
field in the option dialog, which contains both 1s, 2s and 3s for
another few months. Bad me, bad! Thank you, new capybara, for telling
me!

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>