extinfo: Stop hiding interesting information
Whenever a check had not been performed, the extinfo would only say
"This {host,service} has not yet been checked, so status information is
not available." Of course, most of the information (such as if it is in
scheduled downtime, if notications are enabled, when the next check is
scheduled) *is* available, quite applicable - in fact, "next scheduled
check" is rarely as interesting to me as it is when none has been
performed yet.  So don't randomly hide the big status box just because
we can.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>