cucumber: Add even more sleep
Sigh, why don't we just implement a hook to sleep for a bit between
every step?

The way that listviews magically refresh without telling you that is
what they're doing is awfully annoying. It is when you're a user, but
it is even more so when you're a binary running an automated script and
trying to figure out where you are.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>