Extinfo: Display check source node
Information about the check source is sometimes useful. This displays that
information in the extinfo page, but not in the listview.

The list view shuoldn't have that column, because it should be possible to add
as a custom column. It isn't possible to do nice ATM though, because custom
columns can't handle extra fields implemented in the ORM layer, but just native
livestatus fields right now.

Resolves #6762

Signed-off-by: Max Sikström <msikstrom@op5.com>