List View custom columns: rewrite variable resolution
Because the some variables doesn't exist in the livestatus strucutre, but
should be available anyway, we can't depend on the structure definition to
resolve variable availability. That means, we need to rewrite the variable
resolution.

That also means, we need to resolve some variable availabilities on the only
place where possible; in PHP. That means some more logic to the ajax reception
part of listview.

Resolves #7652

Signed-off-by: Max Sikström <msikstrom@op5.com>