Custom Columns: Implement link( rel url, content ) method
To be able to create links from custom columns to other parts of the system,
the link method is nice, because you don't want to be dependent on the address
to monitor to create a link.

Use the syntax link( relative url, content ), where relative url points to
what's after index.php/ in the URL. content is a html string. Result is a
string, where content is wrapped within an <a>-tag

Resolves probably last part of #7102

Signed-off-by: Max Sikström <msikstrom@op5.com>