reports: Make object selection from a generated report possible
There's no real reason not to allow it (my only reason for not including
it sooner was "it wasn't there before"), and this meant that generated
reports needed a hidden version of the selected object list, which due
to the generic nature of these forms caused the hidden list to override
whatever objects you selected yourself. Darn.

So, make the form completely identical between the setup and the report
editing - no special cases, no problem.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>