reports: Remove scheduling info from report options
Scheduling info is not part of a report - it's another object entirely,
with a many-to-one relation against actual reports. Putting it in the
report option was a stupid idea, and I should refrain from doing stupid
things like that in the future.

Luckily, it was sufficiently misplaced to make this aspect virtually
unused, so it was easy to strip out. Lucky me!

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>