report_options: Change property loading, again
I feel it's becoming less convoluted rather than more, so I'm hoping
that's a good thing.

We should not load from DB any option that is
 * already overridden by the user (same as before)
 * invalid in the options array (changed - we never skipped invalid keys
before)
 * set in the options array to something that is not default (should be
the same as the first condition)
 * a bool, where the argument array is "long-ish"
There are fewer nested parenthesis than before, so probably a good
thing.

Where we special case the selected objects, we only do something with
them when there are any, and there is a report type given - otherwise,
weird stacktraces where we try to set "report_type" to an array show up,
which causes crashes.

Finally, when a given report id is invalid, we must immediately get rid
of it, or Bad Things May Happen when other components assume that the
report is, in fact, loaded from database.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>