report_options: Represent bools as numbers
Because when they're represented as bools, they're unprintable and hard
to read or write to/from HTML without casting them to ints, and hard to
represent in a database without casting them to/from ints.

With this change, false is also no longer a valid return value for
getIndex, meaning I get to throw away a bit of weird hacks in
Op5report_options.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>