command: When using fancypants English, do it right or not at all
This has been bugging me for way too long...

"Whence" is silly-old-fashioned English meaning "where from". That means
you don't return "from whence you came", you return "whence you came".
Yes, Shakespeare got away with the redundant "from" from time to time,
but it's bad form nonetheless - Shakespeare got away with all sorts of
nonsense.

But if you're going to do the Shakespeare era English, you completely
failed with the pronouns. Because back then, the modern-day "you" was
not utterly fucked up from centuries of evil Latin/French over-use of
the royal you - the objective singular form was "thee" (to complement
the nominative singular "thou", the nominative plural "ye", and the
objective plural - wait for it! - "you"), so the sentence would be
"Return whence thee came". If, indeed, "return" and "came" were proper
terms back then.

But then we'd need to rewrite every other string to make it as archaic
and hard to understand as this one, and I'm not really good enough at
middle English to do so, so I'm just gonna write "from where" and be
done with it.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>