command: Don't rewrite all submitted commands quite as much
If we're gonna keep multi_command separate from submit, at least have
multi_command paper over the differences once, so submit doesn't have
to care at all. While we're at it, stop putting method-local variables
in the class, and validate input consistently. And look, the
inconsistent validation meant we forgot to check for the presence of a
required input!

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>