helpers/nacoma: Solidify, add check for if service is hostgroup service
Use escapeshellarg() instead of '"'.$arg.'"', because that's the way you
do it. Pay attention to exit codes, and check that it's zero. Use exec to
capture more than one line of output. Don't throw output in /dev/null,
and don't specify a broken second user flag.

Check if service belongs to a hostgroup. Nacoma will auto-resolve the
service when you specify the host name, but it will return it with the
hostgroup name instead of the hostname if that's how it's stored, so
check for the presence of that.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>