cucumber: Fix problem where tests ran too fast
When capybara runs too fast, it would try to click the "Edit report
settings" after clicking "Show report", but before waiting for the
report to reload. As it wasn't visible at that time, things failed.

Have it look at the report the way Kylie would have wanted it to: slow.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>