sass: Fix scrolling and menus in IE, for real this time
Apparently, filter transform is not compatible with position:fixed. In
theory, this discovery should have taken me about 1 man year, but
luckily, here in the 21th century, I cheated and found that out at
http://stackoverflow.com/questions/5664957/css-perhaps-with-compass-cross-browser-gradient

Instead, give IE a fallback to a fixed color for the columns that should
have gradients. This seems to actually work.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>