Auth driver LDAP: add disconnect possibility
Using the Auth Driver LDAP in tests potentially opens a lot of connections,
which is quite easy to keep track of and discconnect during testing.

But the problem is that the disconnect feature of LDAP wasn't exposed, so
expose it. Primarly for testing.

If the feature is intesting for other usage, that method should be exposed
through the abstract driver class too.

(This commit is imported from op5lib)

Signed-off-by: Max Sikström <msikstrom@op5.com>