report_options: NULL is a better "not found" message than false
That is, bools are once again TRUE or FALSE rather than 1 or 0, and when
something isn't found, we instead get a NULL.

The old way was problematic, because when loading saved reports, there's
code that checks if the set value is identical to the default - when the
defaults are all bools, and the set value can't be a bool, that code
won't ever load the saved value.

This meant that in op5reports, the "include summary report" checkbox was
never selected when loading a saved report - after this patch and a
corresponding change in op5report, it will be.

Looking for "=== false" and "!== false" didn't reveal any code that
actually cared about this - I personally only think I've ever used
isset($option['value']) or if($option['value']) so I think we're good.
And, hey, there are tests!

This rewrites e150921 the way chelmertz told me was the best way. Ok,
chelmertz, you win - you knew how to fix this code in a way that didn't
break everything else, and I didn't.

Change-Id: Ib75fdc8a8894102a623f7f2c6230698c172b29b1
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>