Nagios External commands uses *group_name instead of *group attribute
When requesting commands for host and service groups, ninja took the GET
variable hostgroup and servicegroup and put to nagios command field
hostgroup_name and servicegroup_name

The command processing has changed to match nagios commands (afaik), which
means the attribute to the command page is now called host/servicegroup_name.
Make the extinfo page to trigger commands for groups match that.

Resolves #7911

Change-Id: Id39c2dae1af967e485f272850e3efa01bec3806f
Signed-off-by: Max Sikström <msikstrom@op5.com>