Livestatus auth: Downtime table is called downtimes
By some reason, the auth-criteria for donwtime table was set to
match on name "downtime", not "downtimes", which means the auth
criteria wasn't for limited users listing downtimes.

This patch simply changes "downtime" to "downtimes", which means it
matches the table name "downtimes". What the table actually is called
can be confirmed using the table editor in ninja, and it shows with
an s

Resolves #7956

(This commit is imported from op5lib)

Change-Id: I5f23288f46e2289be14cfc197531cbef40c030d8
Signed-off-by: Max Sikstrom <msikstrom@op5.com>