Make op5lib more testable, add hooks to isolate parts for unit testing.
This forces includes to be from the same source, and not use mixed installed
and non-installed files. Much more deterministic when implementing tests.

This commit also fixes some tweaks with linebreaks in the livestatus interface
to not be that hard on linebreaks at the end of the filter.

(This commit is imported from op5lib)

Signed-off-by: Max Sikstrom <msikstrom@op5.com>