Object store: make a central storage for singletons to make testable
To make unit tests, the parts of op5lib should be possible to isolate
and mock.

This patch makes all singletons (except log) to be stored centrally,
so they can be mocked in that lib.

This is the first step of also making the log use that singleton
storage, but log needs to be rewritten, which makes it nessecary to
increase the test suite for the logging

(This commit is imported from op5lib)

Signed-off-by: Max Sikstrom <msikstrom@op5.com>