orm: Make it possible to describe relations between SQL tables
Only tested on two trapper columns so far... Given one table
"foos" => array(
	"class" => "Foo",
	"source" => "SQL",
	"table" => "foo",
	"key" => array("col1", "col2"),
	"structure" => array(
		"col1" => "string",
		"col2" => "string"
	)
)
(everything is still the same as today), you can create a relation to it
with:
"bars" => array(
	"class" => "Bar",
	"source" => "SQL",
	"table" => "bar",
	"key" => array("id"),
	"structure" => array(
		"id" => "int",
		"acol1" => "string",
		"acol2" => "string",
		"afoo" => array("Foo", "afoo")
	),
	"relations" => array(
		array(array("acol1", "acol2"), "foos", "afoo")
	)
)

The bit in the structure,
	"afoo" => array("Foo", "afoo")
creates a association between the two tables. Associations work the same
way they do with a livestatus backend, except the ORM layer expects the
data to be available, but it doesn't know how to ask for it. Hence the
really new bit,
	"relations" => array(
		array(array("acol1", "acol2"), "foos", "afoo")
	)
That is, an array of relations, where each relation consists of the
key(s) used for the join, the name of the other table in the structure
(this implies you can't join between modules), and finally an argument
saying which property in the structure we expect this relation to
populate.

With this in place, you can ask a $bar to get_foo(), and have it return
the foo proprties.

Everything is always joined by default, so the performance is terrible
compared to a real SQL ORM, you cannot get all bars that belong to a
foo, and you can not do many-to-many relations. It's a start, or
something.

(This commit is imported from op5lib)

Change-Id: If55f9a33150ac382ee2136513ba1a366af36161b
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>