Log: Refactor to only use a single singleton, make mocking and testing easier
To mock the log module, the objstore is great to use. To make
objstore easy to use, we need a single instance of the log module.
This patch makes the log available as a singleton, and use a simple
wrapper to augment the returned instance with a namespace parameter

(This commit is imported from op5lib)

Signed-off-by: Max Sikstrom <msikstrom@op5.com>