ormobjectgenerator: array_has_key checks if key is set, isset doesn't
This is somewhat surprising. Let's assume your database returns
NULL. Databases (that aren't livestatus) sometimes may. Isset may appear
to be a method to check that a variable/key is set, but it isn't - it's
a method to check if a variable/key is set to a non-NULL value. Using
array_has_key returns whether the key is actually set, which means that
I can return NULL from the database, without the ORM layer removing the
column, in which case the listview would simply render the next column
in this column's place, which looks strange and is silly.

This also contains an associated typecast for string type columns - I
expect strings to *be* strings.

(This commit is imported from op5lib)

Change-Id: I508821648129b8196a9663c251957e8c95cd53d1
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>