op5Auth::factory - reimplement, but deprecated
To many places use the op5Auth::factory-method. There is better ways
of doing this, for example php has a construct called "new
op5Auth()". But mostly, what is expected is to unload op5auth, and
then create a new instance. It's mostly done in tests.

(This commit is imported from op5lib)

Signed-off-by: Max Sikstrom <msikstrom@op5.com>