lsfiltermain: When an update breaks, at least log it
I don't want to have to figure out how to make it appear beautifully in
the UI, but to take an exception and throw it away completely, while
aborting the update cycle, is the opposite of useful. I've been sitting
for hours now, wondering where the frell my updates went.

Change-Id: Ic80fc3a198abc00be506a32281ea878c35f8994b
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>