quicklink: Allow more characters in URI and title
It's bad to concatenate random strings with HTML, especially if we don't
even escape the quotation signs. It smells like XSS.

Instead, we manipulate DOM nodes' attributes and stop dealing with
strings everywhere. I just don't know how to do it for the icons which
sadly consists both of the .icon-16 *and* the .x16-id classes as
selectors. I guess data-icon was too complicated (or anti cross
browser).

This fixes #7901.

Change-Id: Ide92ce2c3aa2939b83c959c6785d3befad6f0cdd
Signed-off-by: Carl Helmertz <chelmertz@op5.com>