Big UI modifying commit, updated the box-model of ninja to not use as many fixed elements as they just make everything fail.
Added a toolbar controller that currently only the listview
uses.

Added a notification facility in both PHP and JavaScript,
the PHP one simply provides notification data to the JavaScript.

Updated many small margins and paddings, nothing should and longer
use fixed positioning except for things that should (notifications).

Note that this commit is sure to break nomething, be it tests or
functionality, holler at me!

Oh, and also the notifications can handle dialog and so I added
the first part of making ajax commands without page reloads.

The next part will be up soon but is currently in another branch!

Signed-off-by: Tobias Sjondin <tsjondin@op5.com>