config: Force quotes when dumping config
Welcome, welcome everybody! There's still room here in the front!

Today we bring you another episode of our beloved Loose Typing Quiz:
let's guess what happens to an object's type (and value!) as PHP gets
its hands on it!

We'll start off with an easy one: what is the type and value of $a at
the end?
  $a = "123";
  $b = array($a => 0);
  $a = key($b);
The correct answer is, of course, "integer" and "123". As the PHP manual
points out[1], "A key may be either an integer or a string. If a key is
the standard representation of an integer, it will be interpreted as
such", PHP will cast strings to integers whenever it can.

Congratulations to everyone who got it right!

Now, what is the type and value of the following?
 $a = "0123";
 $b = array($a => 0);
 $a = key($b);
The correct answer is "string" and "0123", because the leading zero is
not the "standard representation of an integer" - once again quoting the
dangling quotes from [1], it goes on with "(i.e. "8" will be interpreted
as 8, while "08" will be interpreted as "08")"

So, it's crazy, but not overly insanely so. Assuming, of course, that
"standard representation of an integer" is sanely defined. I wonder if
it is...

Let's move on from core PHP to the spyc library we use here: what is the
type and value of $a in the following:
yaml:
 username: 1
php:
 $b = Spyc::YAMLLoad($yaml_file);
 $a = $b['username']
The correct answer is "integer" and "1". Why? Well, what would you pick?
Keepin' it loose, yo!

And now for something a bit harder: what is the type and value of $a in
the following?
yaml:
 username: 01
php:
 $b = Spyc::YAMLLoad($yaml_file);
 $a = $b['username']
The correct answer is… *drumroll* "float" and "1".

Yeeeeeah....

Hey, at least it's better than an empty array or an SQLResult object,
right? Only barely so, but still!

This commit forces quotes whenever, wherever - because my quotes don't
lie, and thus
 username: "01"
will really be interprented the way you'd expect. This fixes a problem
with users with leading zeroes in their names, solving bug #8052. The
problem is easier to test in nacoma, so that's where a test is added in
change I8ad708303acc0f87180fbfb8df1c57ca7da26dca.

[1] the exact quote I use was copied from the manual into
http://stackoverflow.com/questions/4100488/a-numeric-string-as-array-key-in-php
but the manual appears to have removed the sentence. You can still find
example #8 containing:
 $switching = array(
 [...]
 '8'  => 2, // key = 8 (integer!)
 [...]
 );

(This commit is imported from op5lib)

Change-Id: I7a14667e0969a5e661a0cfc9a54bb1bd7c748cfc
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>