listview: Enable modules to add extra javascript
If a module installed a new listview table, and said listview table had
actions or similar that needed extra javascript to work, then tough
luck.

This makes it possible to easily specify in your manifest that you
wish listviews to load your additional javascript. It will be loaded for
all listviews always, because it's easier that way and listviews don't
reload when changing table.

Change-Id: If659cb93a3f321eb27cfe929b3acc2c5ca8737d6
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>