listview: API-ify buttons, dropdowns
That is, let each module specify as few or as many buttons in the top
right of a list view as it wants to, and to specify for which tables it
is relevant.

For ninja, this would be useful to let the downtime listview add a "Add
new downtime" button and possibly dropdown.

For trapper, being able to not only add buttons, but also dropdowns, is
absolutely essential. I expect something similar to be true for
logserver, whenever that becomes a listview.

Change-Id: I2523e6d7360ad0a25ed7348bc56884b6caca4141
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>