lsfilter: Make button part of listview a link
The stats information to the left is there and styled beatifully,
whether we use text or not. The buttons on the right looks good enough,
but if you put text inside the button, weird things happen (is that even
supposed to work?), and we'd need to double up on styling to make both
links and buttons look the same. Just don't.

Change-Id: Ie9a071b799b00043a065efbd70c571aa35a0ab0b
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>