reports.sass: Add focus & hover looks for buttons
Personally, I find it annoying not to know which <select> that's focused
(outline: 0) so I removed it and let it default to the browser's builtin
styling. That's probably a lot more useful for more people than me.

Also a bit of cleanup since there were so many "input" at the top level

Change-Id: Iba1fa3690a174414c8203689fdcedc04f32768df
Signed-off-by: Carl Helmertz <chelmertz@op5.com>