js: Downgrade jquery selectboxes
Nacoma carries an older, unminimized v2.2.3. As it's not minimized,
we've been able to apply bugfixes to it.

Ninja carried a newer, minimized v2.2.4. As it's minimized, we've not
been able to apply any bugfixes to it.

It's incredibly stupid not to sync up these two. As nacoma's selectboxes
are under heavier use than ninja's, my gut tells me to stick with that.

Change-Id: Ia066a9bff9cc5913cb8162e6f6f40dfc4cf4ce60
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>