reports/js: Lower loading saved report complexity
To initialize, we used to first traverse the objects you could access in
linear time and put them in the available objects multiselect. After
that, an extra pass would actually choose the elements by traversing the
javascript list that was there all along with the elements that should
be pre-selected, look for an object with that name in the long and heavy
unselected select, and, if found, would select the element from the
unselected select, and then finally move all the selected options from
the unselected select to the selected select, and finally, sort the
selected selects, because even though the list of objects comes
preselected from livestatus, we decided to select them in the order
they're stored in the database, which isn't sorted.

This patch reduces this to only doing one pass - when iterating over the
list of available objects, immediately put the ones that should go into
the selected select to their final location, and skip all that other
work.

Change-Id: I6438ab1f2368cff59c012236b693aceac90ad0e9
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>