reports: Remove name from unselected multiselects
The "name" attribute is what makes the browser want to submit them -
from http://www.w3.org/TR/html401/interact/forms.html we can see that 'A
successful control is "valid" for submission. [...] A successful control
must be defined within a FORM element and must have a control name.'
Now, moving the unselected control outside of the form element sounds
painful, so let's just delete the control name.

This removes the need to traverse the frequently gi-normous select,
meaning you can generate a report quickly without having to enter a
regex that excludes the elements in the unselected list (as that is
itself slow).

Change-Id: I1a0f41fd5e9830a83335b921adf25c4c54d9da38
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>