reports: Make it possible to include long output in summaries
And for the record, summaries includes alert histories.

It is off by default, and must be activated per alert history/recent
alert view. We should update the defaults in tandem with #7729 for great
UI consistency.

However, as a special case, after a bit of discussion with fnord, I've
decided to enable it by default when you are at an extinfo page, and go
to that very same object's alert history - as you'll probably have very
similar output for all lines, the "would mess up my layout" argument
doesn't apply.

This fixes bug #7993.

Change-Id: I120e19a2d46a5a912da6bd0eac6f26e7687549a2
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>