search: Make search work for data in long_plugin_output also
Previously, we only searched plugin_output for "si" and other searches,
and so we only searched through the first line of plugin output.

Many plugins, some of which we ship default with the product, output
important data on lines 2 and beyond. This was, then, a bad idea.

It remains to be seen whether or not searching through long_plugin_output
is a bad idea too - mainly the worry would be that you could suddenly be
searching through much much more data than before, but.. meh, nobody likes
a coward.

(also, my (limited) tests do not show a measurable performance impact)

Signed-off-by: Mattias Bergsten <mattias.bergsten@op5.com>