AuthDriver: Header auth, read authentication from http headers
When protecting the op5 system behind an authenticating proxy, and
the proxy sends out authorization information as HTTP-headers, and
you can trust that it isn't accessed directly from a client, which
can fake the headers, this module can be used to read those
credentials and create a user.

Make sure that the op5 system is well proceted behind a firewall +
the authenticating proxy before enabling this though.

This also includes unit tests, with a mocked HTTP header setup

(This commit is imported from op5lib)

Change-Id: Ife3ea70ccadec06cf47ee3b640b52a1dca5c5d72
Tests: #8104
Resolves: #8104
Signed-off-by: Max Sikstrom <msikstrom@op5.com>