AuthDriver_Header: Use , as default group delimiter
According to RFC 2616, section 4.2:

START QUOTE

Multiple message-header fields with the same field-name MAY be
present in a message if and only if the entire field-value for that
header field is defined as a comma-separated list [i.e., #(values)].
It MUST be possible to combine the multiple header fields into one
"field-name: field-value" pair, without changing the semantics of the
message, by appending each subsequent field-value to the first, each
separated by a comma. The order in which header fields with the same
field-name are received is therefore significant to the
interpretation of the combined field value, and thus a proxy MUST NOT
change the order of these field values when a message is forwarded.

END QUOTE

The default delimiter shuold probably be , instead of space. It is
still configurable, but you need to explicitly configure it to space
if you want to.

(This commit is imported from op5lib)

Change-Id: I832df456228d9f6de1c26d6e21e411995b6417cf
Signed-off-by: Max Sikstrom <msikstrom@op5.com>