Replace qtip 1.0.0-rc3 with qtip 2.2.0
According to [1]: "How to destroy a qtip when the original element is
removed

Move over to qTip2, which implements a $.remove() override identical to
that in jQuery UI, which triggers the removal of any bound qTips when an
element that has qTip or any parent elements are removed."

The qTip2 version that I added is "2.2.0 stable" with the included
styles: core styles, basic colour styles. The additional features:
"viewport adjustment" and no additional libraries. (I realize that his
download wizard can change at any moment but this is as good as I can
describe it.)

qtip2 has a completely new syntax which I tried to take in account when
modifying our code. To ease this, template_head.php now defines
window.qtip_default which rids most of the copy & paste.

This fixes #6989.

  1: http://craigsworks.com/projects/forums/showthread.php?tid=1935

Change-Id: Ifef4092a9901cd7f98503c84d1c25ea9445081dc
Signed-off-by: Carl Helmertz <chelmertz@op5.com>