showlog: Use new showlog protocol for limited users
This fixes the ninja half of #7960.

Despite PHP's arrays having O(n) lookup[1], and the host array can grow
quite long, I'm betting that "compressing" the size of the string
leaving PHP will create better performance.

[1] http://php.net/manual/en/language.types.array.php: "An array in PHP
is actually an ordered map. A map is a type that associates values to
keys. This type is optimized for several different uses; it can be
treated as an array, list (vector), hash table"
In other words, PHP arrays are *like* hash tables, in much the same way
as how air is *like* water - both ordered maps and hashes have keys that
maps to values, and both air and water is used by animals to extract
life-essential oxygen. Which is why it'll be completely OK if I push the
author of that text into the ocean and hold their head under water if I
ever meet them.

Change-Id: Ie83a7bb5111313fada7d04ac7630b598f30a75d5
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>