restore: Tell restore which files to extract
Append file list to restore command as a security measure making sure no
unwanted files are extracted from the tar archive.

Intersection is done to avoid "Not found in archive" errors if user wants to
restore from an old archive or if the file in the list wasn't available when the
backup was taken.

This together with 1369c60269a7df3fee8597aaa99f9bd341f749a1 fixes #7943

Signed-off-by: Henrik Smederod <hsmederod@op5.com>