AuthDriver_Header: Add debug logging, don't accept unauthorized users
Previously, when doing autologin, the user was accepted as long as it
was a valid user object, but it didn't need any credentials. That was
fine, because all previous auto_login-drivers didn't support groups.
Fix so the user needs some kind of permissions to be accepted by
auto_login.

Also, add some (noisy) debug logging, so it logs each page load with
which users is authenticated, and which groups. This will fill the log
files quite fast if activated in a long time.

Resolves #8104 and adds backend support to identify that header auth
can't log out.

(This commit is imported from op5lib)

Change-Id: I77518134e0d52a77d8a73c7ce70aae4377d4f756
Signed-off-by: Max Sikstrom <msikstrom@op5.com>