Janitor away some showlog code
Removing 'functionality' in form of a long "is the start date later than
the end date?" check. That check referred to undefined variables and
choked. Now, if you try to do that, you get an empty result page. The
user would have to trick the builtin datepickers for that though.

This fixes #6961.

Change-Id: Ife4477ed54ee1f2059d0261ba5fa421cca15e73c
Signed-off-by: Carl Helmertz <chelmertz@op5.com>