Make it possible to add global notifications through event handlers
The Event system in Kohana + hooks is really powerful to be able to
drop in modifications to pages. So make it possible to drop in a module
with hooks, which can add global notificaitons in the
system.post_controller_constructor and system.post_controller events.

Unfourtunatly Kohana missed to send the controller through to those
events as data, which made them unavailable for the event handler.

Change-Id: Ie0c321ce4d897bb399cf2e93e02d4325e5a90482
Signed-off-by: Max Sikstrom <msikstrom@op5.com>