recurring_downtime: Put into listview, standardize
The list of recurring downtimes belong as a listview, not as a custom,
hardcoded list. Listviews can be filtered, sorted, and have a kind of
flexibility that is really neat.

Change database layout - we have a very clear schema that we expect the
model to return, we just didn't store it as database columns. While the
ORM layer can fake columns, you can't use them to sort, which made the
listview quite boring until the layout was changed.

Straighten out the difference between a flexible downtime and a
triggered downtime. The original author of this code appears to have
been under the impression that a downtime can be fixed or triggered,
while the triggeredness is completely orthogonal to whether the downtime
is fixed or flexible. A user should be allowed to enter a start time and
an end time for a fixed downtime, while they should also be allowed to
enter a duration for a flexible downtime. Triggered recurring downtimes
make no sense at all (unless it would be possible to trigger a downtime
based on another schema, which is far beyond what is actually possible),
so that is completely deleted. This fixes bug #7641 as well as #7637.

While we're at it, implement permissions properly - it used to be you
were authorized to see and edit everything if you had host_view_all, and
was only allowed to delete if you had system_commands. This looks like a
straight-up translation of the poor state of our permissions in
monitor 5, but we now have real, nice, useful permissions, so let's use
them.

The JQuery UI Tab interface that the recurring downtime page used is
completely unheard of outside of this controller these days - it used to
be used in the report scheduling, but that is now its on thing, so the
tab interface looked foreign (not least because everything controlled by
a tab had a tiny bit smaller font). Use a regular notify bar to switch
between the two.

As a bit of an "experiment", the object selector is converted from
having the same four copies of the object multi-select that all all
other reports have, into having one single multi-select.

Change-Id: I4cf0db91fd1e433c2710dc8fafc71b1e2cbc85fd
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>