Clean up and move away legacy stuff
Those classes isn't used anywhere (as it appears) within ninja, so move
them out of harms way.

They might be used in widgets however, so it would be great to look
through the list, and test the system without those classes, so we can
remove them later.

The nagios_auth_model can't be used however, because it simply doesn't
work, and if something still uses it, it's better if it breaks hard.

The old_comment_model class was previously used by two
services-widgets. The only method that was used in the model is now
moved to that widget, and that widget itself is legacy. So
old_comment_model is safe to throw away.

The db model isn't used anywhere within monitor, and I can't see the
use of it anywhere else. If someone uses it in widgets, they should
probably copy the functions from a previous revision to their widgets
itself. But if they use it, it's also probably because they use the
legacy merlin database as backend, which doesn't work with this version
of monitor and ninja.

Change-Id: I43583375bb1ed9acebc872fb6bead24cb3fefe51
Signed-off-by: Max Sikstrom <msikstrom@op5.com>